Add TODO list of useful features to add
I added a TODO list at the end of contents.md because I think although
your tutorial is fantastic, it is missing some really critical parts
which must be added to make it complete, namely: The "#" Goto operator
(Ctrl + ;), how to use Projects and Workspaces (critical!--in order to
really use Sublime properly) and a list of Packages I think are worth
mentioning. Of the packages I list, I recognize that many of them are
based on personal preference or taste. However, 3 critical ones that I
think you *must* mention are Print to Html, ExportHtml, and Compare
Side-By-Side. The first two add printing capability to Sublime Text--
which many people consider to be an essential feature and won't want to
switch over to Sublime if they don't have printing ability. As for
Compare Side-By-Side, I think it's just too useful not to include. I use
it all the time when I need to get a better feel for what's going on
between a couple files. For noobs in particular it is waaay easier to
use than `git diff`.

That is all! I hope you will incorporate this commit as a pull request.
At a bare minimum it provides some important links and general info to
get someone who just completed your tutorial started on their own
reading. As a longer term goal it gives you some material and a starting
point to complete this otherwise almost perfect tutorial. Thanks!